+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.16.0
+* Allow Bundler executable to be customized.
+  - Main use-case is to allow adding in the Prebundler plugin, https://github.com/getkuby/kuby-prebundler.
+* Print error message and exit on missing Docker build args.
+  - Ignore with `--ignore-missing-args`.
+
 ## 0.15.0
 * Add an extra parameter to the Docker spec for specifying the URL of the Docker registry index.
   - In most cases, the registry and registry index URLs will be the same. However Docker Hub, the default registry, uses index.docker.io for API requests (catalog, tags, etc) but only allows pushes to docker.io.